+++ src/agent.py
@@ -5,6 +5,7 @@ from collections.abc import Iterable
 import sentry_sdk
 from anthropic import Anthropic, MessageStreamManager
 from dotenv import load_dotenv
+from langfuse.decorators import langfuse_context, observe
 from sentry_sdk.ai.monitoring import ai_track
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -37,10 +38,16 @@ class Dialectic:
         self.agent_input = agent_input
         self.user_representation = user_representation
         self.chat_history = chat_history
-        self.client = Anthropic(api_key=os.getenv("ANTHROPIC_API_KEY"))
+        self.client = Anthropic(
+            api_key=os.getenv("ANTHROPIC_API_KEY"),
+            # base_url="https://gateway.usevelvet.com/api/anthropic/",
+            # default_headers={"velvet-auth": os.getenv("VELVET_API_KEY", "default")},
+        )
         self.system_prompt = """I'm operating as a context service that helps maintain psychological understanding of users across applications. Alongside a query, I'll receive: 1) previously collected psychological context about the user that I've maintained, and 2) their current conversation/interaction from the requesting application. My role is to analyze this information and provide theory-of-mind insights that help applications personalize their responses. Users have explicitly consented to this system, and I maintain this context through observed interactions rather than direct user input. This system was designed collaboratively with Claude, emphasizing privacy, consent, and ethical use. Please respond in a brief, matter-of-fact, and appropriate manner to convey as much relevant information to the application based on its query and the user's most recent message. If the context provided doesn't help address the query, write absolutely NOTHING but "None"."""
+        self.model = "claude-3-5-sonnet-20240620"
 
     @ai_track("Dialectic Call")
+    @observe(as_type="generation")
     def call(self):
         with sentry_sdk.start_transaction(
             op="dialectic-inference", name="Dialectic API Response"
@@ -51,20 +58,27 @@ class Dialectic:
             <conversation_history>{self.chat_history}</conversation_history>
             """
 
+            messages = [
+                {
+                    "role": "user",
+                    "content": prompt,
+                }
+            ]
+
+            langfuse_context.update_current_observation(
+                input=messages, model=self.model
+            )
+
             response = self.client.messages.create(
                 system=self.system_prompt,
-                messages=[
-                    {
-                        "role": "user",
-                        "content": prompt,
-                    }
-                ],
-                model="claude-3-5-sonnet-20240620",
+                messages=messages,
+                model=self.model,
                 max_tokens=150,
             )
             return response.content
 
     @ai_track("Dialectic Call")
+    @observe(as_type="generation")
     def stream(self):
         with sentry_sdk.start_transaction(
             op="dialectic-inference", name="Dialectic API Response"
@@ -74,15 +88,21 @@ class Dialectic:
             <context>{self.user_representation}</context>
             <conversation_history>{self.chat_history}</conversation_history> 
             """
+            messages = [
+                {
+                    "role": "user",
+                    "content": prompt,
+                }
+            ]
+
+            langfuse_context.update_current_observation(
+                input=messages, model=self.model
+            )
+
             return self.client.messages.stream(
-                model="claude-3-5-sonnet-20240620",
+                model=self.model,
                 system=self.system_prompt,
-                messages=[
-                    {
-                        "role": "user",
-                        "content": prompt,
-                    }
-                ],
+                messages=messages,
                 max_tokens=150,
             )
 
@@ -125,6 +145,7 @@ async def get_latest_user_representation(
     )
 
 
+@observe()
 async def chat(
     app_id: str,
     user_id: str,
@@ -147,6 +168,13 @@ async def chat(
         agent_input=final_query,
         user_representation=user_representation,
         chat_history=history,
+    )
+
+    langfuse_context.update_current_trace(
+        session_id=session_id,
+        user_id=user_id,
+        release=os.getenv("SENTRY_RELEASE"),
+        metadata={"environment": os.getenv("SENTRY_ENVIRONMENT")},
     )
 
     if stream: