+++ src/deriver/__main__.py
@@ -7,4 +7,3 @@ from .queue import main
 if __name__ == "__main__":
     asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
     asyncio.run(main())
-